# B2B_API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: "Dream ERP B2B Integration API"
  description: "JSON-EDI Protocol for B2B Purchase Orders, Invoices, and Ledger Integration"
  version: "1.0.0"
  contact:
    name: "Dream ERP Platform"
    email: "integration@dream-erp.com"
servers:
  - url: "https://api.dream-erp.com/v1"
    description: "Production API"
  - url: "https://staging-api.dream-erp.com/v1"
    description: "Staging (Testing) API"
security:
  - BearerAuth: []
  - OAuth2: [read, write]
...
paths:
...
  /b2b/webhooks:
    post:
      tags:
        - "Webhooks"
      summary: "Register webhook URL for real-time notifications"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegistration'
      responses:
        '201':
          description: "Webhook registered successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "OAuth2 JWT token from Keycloak"
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: "https://auth.dream-erp.com/oauth2/authorize"
          tokenUrl: "https://auth.dream-erp.com/oauth2/token"
          scopes:
            read: "Read purchase orders and invoices"
            write: "Create and acknowledge purchase orders"
  schemas:
...
    WebhookRegistration:
      type: object
      required: [webhook_url, events]
      properties:
        webhook_url:
          type: string
          format: uri
          example: "https://partner.com/api/webhooks/erp"
        events:
          type: array
          items:
            type: string
            enum: [po.received, po.acknowledged, order.shipped, invoice.issued, payment.received]
        active:
          type: boolean
          default: true
    WebhookResponse:
      type: object
      properties:
        webhook_id:
          type: string
        webhook_url:
          type: string
        status:
          type: string
        registered_at:
          type: string
          format: date-time
...